app/api/requirements.txt
geopandas==0.14.1
numpy==1.26.2
pandas==2.1.3
fastapi==0.104.1
tortoise-orm==0.20.0
uvicorn[standard]==0.24.0.post1
aerich==0.7.2
